+++ lib/rig/distributed_set.ex
@@ -6,6 +6,7 @@ defmodule RIG.DistributedSet do
   use GenServer
   import Ex2ms
 
+  alias RigMetrics.DistributedSetMetrics
   alias Timex
   alias UUID
 
@@ -168,12 +169,15 @@ defmodule RIG.DistributedSet do
   def handle_call(
         {:add_from_local, {_, uuid, _, _} = record},
         _from,
-        %{pg2_group: pg2_group, ets_table: ets_table, last_record_id: last_record_id} = state
+        %{pg2_group: pg2_group, ets_table: ets_table, last_record_id: last_record_id, name: name} =
+          state
       ) do
     ets_table
     |> save(record)
     |> broadcast_update(pg2_group, last_record_id)
 
+    # increase Prometheus metrics (total, current) number of items in this set
+    DistributedSetMetrics.add_item(name)
     {:reply, :ok, %{state | last_record_id: uuid}}
   end
 
@@ -181,7 +185,7 @@ defmodule RIG.DistributedSet do
   @impl GenServer
   def handle_info(
         {:add_from_remote, {_, incoming_record_id, _, _} = record, remote_previous_record_id},
-        %{ets_table: ets_table, last_record_id: local_last_record_id} = state
+        %{ets_table: ets_table, last_record_id: local_last_record_id, name: name} = state
       ) do
     save(ets_table, record)
 
@@ -208,6 +212,8 @@ defmodule RIG.DistributedSet do
           local_last_record_id
       end
 
+    # increase Prometheus metrics (total, current) number of items in this set
+    DistributedSetMetrics.add_item(name)
     {:noreply, %{state | last_record_id: synced_record_id}}
   end
 
@@ -241,15 +247,15 @@ defmodule RIG.DistributedSet do
 
   # Remove expired records.
   @impl GenServer
-  def handle_info(:cleanup, %{ets_table: ets_table} = state) do
-    remove_expired_records(ets_table)
+  def handle_info(:cleanup, %{ets_table: ets_table, name: name} = state) do
+    remove_expired_records(ets_table, name)
     Process.send_after(self(), :cleanup, @cleanup_interval_s * 1_000)
     {:noreply, state}
   end
 
   # private
 
-  defp remove_expired_records(ets_table) do
+  defp remove_expired_records(ets_table, name) do
     now = Timex.now() |> serialize_datetime!()
 
     match_spec =
@@ -263,6 +269,8 @@ defmodule RIG.DistributedSet do
       Logger.debug(fn -> "Removed #{n_deleted} expired records" end)
     end
 
+    # decrease Prometheus metric (current) number of items in this set
+    DistributedSetMetrics.delete_item(name, n_deleted)
     n_deleted
   end